Put GetCanonicalPath into its own utility class
